format data-mysql

+++ data-mysql/src/main/java/org/adamalang/mysql/frontend/Authorities.java
@@ -18,17 +18,10 @@ import java.util.ArrayList;
 
 public class Authorities {
 
-  public static int createAuthority(DataBase dataBase, int ownerId, String authority)
-      throws Exception {
+  public static int createAuthority(DataBase dataBase, int ownerId, String authority) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("INSERT INTO `")
-              .append(dataBase.databaseName)
-              .append("`.`authorities` (`owner`, `authority`, `keystore`) VALUES (?,?,'{}')")
-              .toString();
-      try (PreparedStatement statement =
-          connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+      String sql = new StringBuilder().append("INSERT INTO `").append(dataBase.databaseName).append("`.`authorities` (`owner`, `authority`, `keystore`) VALUES (?,?,'{}')").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
         statement.setInt(1, ownerId);
         statement.setString(2, authority);
         statement.execute();
@@ -39,17 +32,10 @@ public class Authorities {
     }
   }
 
-  public static void setKeystore(DataBase dataBase, int ownerId, String authority, String keystore)
-      throws Exception {
+  public static void setKeystore(DataBase dataBase, int ownerId, String authority, String keystore) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("UPDATE `")
-              .append(dataBase.databaseName)
-              .append("`.`authorities` SET `keystore`=? WHERE `owner`=? AND authority=?")
-              .toString();
-      try (PreparedStatement statement =
-          connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+      String sql = new StringBuilder().append("UPDATE `").append(dataBase.databaseName).append("`.`authorities` SET `keystore`=? WHERE `owner`=? AND authority=?").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
         statement.setString(1, keystore);
         statement.setInt(2, ownerId);
         statement.setString(3, authority);
@@ -63,41 +49,23 @@ public class Authorities {
   public static ArrayList<String> list(DataBase dataBase, int ownerId) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
       ArrayList<String> results = new ArrayList<>();
-      String sql =
-          new StringBuilder()
-              .append("SELECT `authority` FROM `")
-              .append(dataBase.databaseName)
-              .append("`.`authorities` WHERE owner=")
-              .append(ownerId)
-              .append(" ORDER BY `authority` ASC")
-              .toString();
-      DataBase.walk(
-          connection,
-          (rs) -> {
-            results.add(rs.getString(1));
-          },
-          sql);
+      String sql = new StringBuilder().append("SELECT `authority` FROM `").append(dataBase.databaseName).append("`.`authorities` WHERE owner=").append(ownerId).append(" ORDER BY `authority` ASC").toString();
+      DataBase.walk(connection, (rs) -> {
+        results.add(rs.getString(1));
+      }, sql);
       return results;
     }
   }
 
-  public static void changeOwner(
-      DataBase dataBase, String authority, int oldOwnerId, int newOwnerId) throws Exception {
+  public static void changeOwner(DataBase dataBase, String authority, int oldOwnerId, int newOwnerId) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("UPDATE `")
-              .append(dataBase.databaseName)
-              .append("`.`authorities` SET `owner`=? WHERE `owner`=? AND authority=?")
-              .toString();
-      try (PreparedStatement statement =
-          connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+      String sql = new StringBuilder().append("UPDATE `").append(dataBase.databaseName).append("`.`authorities` SET `owner`=? WHERE `owner`=? AND authority=?").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
         statement.setInt(1, newOwnerId);
         statement.setInt(2, oldOwnerId);
         statement.setString(3, authority);
         if (statement.executeUpdate() != 1) {
-          throw new ErrorCodeException(
-              ErrorCodes.FRONTEND_AUTHORITY_CHANGE_OWNER_NOT_FOUND_OR_INCORRECT);
+          throw new ErrorCodeException(ErrorCodes.FRONTEND_AUTHORITY_CHANGE_OWNER_NOT_FOUND_OR_INCORRECT);
         }
       }
     }
@@ -105,14 +73,8 @@ public class Authorities {
 
   public static String getKeystoreInternal(DataBase dataBase, String authority) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("SELECT `keystore` FROM `")
-              .append(dataBase.databaseName)
-              .append("`.`authorities` WHERE authority=?")
-              .toString();
-      try (PreparedStatement statement =
-               connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+      String sql = new StringBuilder().append("SELECT `keystore` FROM `").append(dataBase.databaseName).append("`.`authorities` WHERE authority=?").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
         statement.setString(1, authority);
         ResultSet rs = statement.executeQuery();
         if (rs.next()) {
@@ -126,14 +88,8 @@ public class Authorities {
 
   public static String getKeystorePublic(DataBase dataBase, int owner, String authority) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("SELECT `keystore` FROM `")
-              .append(dataBase.databaseName)
-              .append("`.`authorities` WHERE authority=? AND `owner`=?")
-              .toString();
-      try (PreparedStatement statement =
-               connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+      String sql = new StringBuilder().append("SELECT `keystore` FROM `").append(dataBase.databaseName).append("`.`authorities` WHERE authority=? AND `owner`=?").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
         statement.setString(1, authority);
         statement.setInt(2, owner);
         ResultSet rs = statement.executeQuery();
@@ -147,18 +103,10 @@ public class Authorities {
   }
 
 
-
-  public static void deleteAuthority(DataBase dataBase, int ownerId, String authority)
-      throws Exception {
+  public static void deleteAuthority(DataBase dataBase, int ownerId, String authority) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("DELETE FROM `")
-              .append(dataBase.databaseName)
-              .append("`.`authorities` WHERE `owner`=? AND authority=?")
-              .toString();
-      try (PreparedStatement statement =
-          connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+      String sql = new StringBuilder().append("DELETE FROM `").append(dataBase.databaseName).append("`.`authorities` WHERE `owner`=? AND authority=?").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
         statement.setInt(1, ownerId);
         statement.setString(2, authority);
         if (statement.executeUpdate() != 1) {
