+++ data-mysql/src/main/java/org/adamalang/mysql/frontend/Spaces.java
@@ -14,19 +14,16 @@ import org.adamalang.common.ErrorCodeException;
 import org.adamalang.mysql.DataBase;
 
 import java.sql.*;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 public class Spaces {
   public static int createSpace(DataBase dataBase, int userId, String space) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sqlTestWater =
-          new StringBuilder()
-              .append("SELECT `owner`, `id` FROM `")
-              .append(dataBase.databaseName)
-              .append("`.`spaces` WHERE `name`=?")
-              .toString();
-      try (PreparedStatement statement =
-          connection.prepareStatement(sqlTestWater, Statement.RETURN_GENERATED_KEYS)) {
+      String sqlTestWater = new StringBuilder().append("SELECT `owner`, `id` FROM `").append(dataBase.databaseName).append("`.`spaces` WHERE `name`=?").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sqlTestWater, Statement.RETURN_GENERATED_KEYS)) {
         statement.setString(1, space);
         ResultSet rs = statement.executeQuery();
         if (rs.next()) {
@@ -35,14 +32,8 @@ public class Spaces {
           }
         }
       }
-      String sql =
-          new StringBuilder()
-              .append("INSERT INTO `")
-              .append(dataBase.databaseName)
-              .append("`.`spaces` (`owner`, `name`, `plan`, `hash`, `billing`) VALUES (?,?,'{}', '', 'free')")
-              .toString();
-      try (PreparedStatement statement =
-          connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+      String sql = new StringBuilder().append("INSERT INTO `").append(dataBase.databaseName).append("`.`spaces` (`owner`, `name`, `plan`, `hash`, `billing`) VALUES (?,?,'{}', '', 'free')").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
         statement.setInt(1, userId);
         statement.setString(2, space);
         statement.execute();
@@ -55,11 +46,7 @@ public class Spaces {
 
   public static Space getSpaceId(DataBase dataBase, String space) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder("SELECT `id`,`owner`,`billing` FROM `")
-              .append(dataBase.databaseName)
-              .append("`.`spaces` WHERE name=?")
-              .toString();
+      String sql = new StringBuilder("SELECT `id`,`owner`,`billing` FROM `").append(dataBase.databaseName).append("`.`spaces` WHERE name=?").toString();
       try (PreparedStatement statement = connection.prepareStatement(sql)) {
         statement.setString(1, space);
         try (ResultSet rs = statement.executeQuery()) {
@@ -67,22 +54,14 @@ public class Spaces {
             Set<Integer> developers = new HashSet<>();
             int owner = rs.getInt(2);
             developers.add(owner);
-            String sqlGrants =
-                new StringBuilder("SELECT `user`,`role` FROM `")
-                    .append(dataBase.databaseName)
-                    .append("`.`grants` WHERE `space`=")
-                    .append(rs.getInt(1))
-                    .toString();
-            DataBase.walk(
-                connection,
-                (g) -> {
-                  switch (g.getInt(2)) {
-                    case 0x01: // Role.Developer
-                      developers.add(g.getInt(1));
-                      break;
-                  }
-                },
-                sqlGrants);
+            String sqlGrants = new StringBuilder("SELECT `user`,`role` FROM `").append(dataBase.databaseName).append("`.`grants` WHERE `space`=").append(rs.getInt(1)).toString();
+            DataBase.walk(connection, (g) -> {
+              switch (g.getInt(2)) {
+                case 0x01: // Role.Developer
+                  developers.add(g.getInt(1));
+                  break;
+              }
+            }, sqlGrants);
             return new Space(rs.getInt(1), owner, rs.getString(3), developers);
           }
           throw new ErrorCodeException(ErrorCodes.FRONTEND_SPACE_DOESNT_EXIST);
@@ -93,16 +72,8 @@ public class Spaces {
 
   public static void setPlan(DataBase dataBase, int spaceId, String plan, String hash) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("UPDATE `")
-              .append(dataBase.databaseName)
-              .append("`.`spaces` SET `plan`=?, `hash`=? WHERE `id`=")
-              .append(spaceId)
-              .append(" LIMIT 1")
-              .toString();
-      try (PreparedStatement statement =
-          connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+      String sql = new StringBuilder().append("UPDATE `").append(dataBase.databaseName).append("`.`spaces` SET `plan`=?, `hash`=? WHERE `id`=").append(spaceId).append(" LIMIT 1").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
         statement.setString(1, plan);
         statement.setString(2, hash);
         statement.execute();
@@ -112,16 +83,8 @@ public class Spaces {
 
   public static void setBilling(DataBase dataBase, int spaceId, String billing) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("UPDATE `")
-              .append(dataBase.databaseName)
-              .append("`.`spaces` SET `billing`=? WHERE `id`=")
-              .append(spaceId)
-              .append(" LIMIT 1")
-              .toString();
-      try (PreparedStatement statement =
-          connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+      String sql = new StringBuilder().append("UPDATE `").append(dataBase.databaseName).append("`.`spaces` SET `billing`=? WHERE `id`=").append(spaceId).append(" LIMIT 1").toString();
+      try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
         statement.setString(1, billing);
         statement.execute();
       }
@@ -130,12 +93,7 @@ public class Spaces {
 
   public static String getPlan(DataBase dataBase, int spaceId) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder("SELECT `plan` FROM `")
-              .append(dataBase.databaseName)
-              .append("`.`spaces` WHERE id=")
-              .append(spaceId)
-              .toString();
+      String sql = new StringBuilder("SELECT `plan` FROM `").append(dataBase.databaseName).append("`.`spaces` WHERE id=").append(spaceId).toString();
       try (PreparedStatement statement = connection.prepareStatement(sql)) {
         try (ResultSet rs = statement.executeQuery()) {
           if (rs.next()) {
@@ -147,23 +105,9 @@ public class Spaces {
     }
   }
 
-  public static class InternalDeploymentPlan {
-    public final String plan;
-    public final String hash;
-
-    public InternalDeploymentPlan(String plan, String hash) {
-      this.plan = plan;
-      this.hash = hash;
-    }
-  }
-
   public static InternalDeploymentPlan getPlanByNameForInternalDeployment(DataBase dataBase, String spaceName) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder("SELECT `plan`, `hash` FROM `")
-              .append(dataBase.databaseName)
-              .append("`.`spaces` WHERE `name`=?")
-              .toString();
+      String sql = new StringBuilder("SELECT `plan`, `hash` FROM `").append(dataBase.databaseName).append("`.`spaces` WHERE `name`=?").toString();
       try (PreparedStatement statement = connection.prepareStatement(sql)) {
         statement.setString(1, spaceName);
         try (ResultSet rs = statement.executeQuery()) {
@@ -176,34 +120,18 @@ public class Spaces {
     }
   }
 
-  public static List<Item> list(DataBase dataBase, int userId, String marker, int limit)
-      throws Exception {
+  public static List<Item> list(DataBase dataBase, int userId, String marker, int limit) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
       // select * from a LEFT OUTER JOIN b on a.a = b.b;
-      String sql =
-          new StringBuilder("SELECT `s`.`name`,`s`.`owner`,`s`.`billing`,`s`.`created` FROM `")
-              .append(dataBase.databaseName) //
-              .append("`.`spaces` as `s` LEFT OUTER JOIN `")
-              .append(dataBase.databaseName)
-              .append("`.`grants` as `g` ON `s`.`id` = `g`.`space`") //
-              .append(" WHERE (`s`.owner=")
-              .append(userId)
-              .append(" OR `g`.`user`=")
-              .append(userId)
-              .append(") AND `s`.`name`>? ORDER BY `s`.`name` ASC LIMIT ")
-              .append(limit)
-              .toString();
+      String sql = new StringBuilder("SELECT `s`.`name`,`s`.`owner`,`s`.`billing`,`s`.`created` FROM `").append(dataBase.databaseName) //
+                                                                                                        .append("`.`spaces` as `s` LEFT OUTER JOIN `").append(dataBase.databaseName).append("`.`grants` as `g` ON `s`.`id` = `g`.`space`") //
+                                                                                                        .append(" WHERE (`s`.owner=").append(userId).append(" OR `g`.`user`=").append(userId).append(") AND `s`.`name`>? ORDER BY `s`.`name` ASC LIMIT ").append(limit).toString();
       try (PreparedStatement statement = connection.prepareStatement(sql)) {
         statement.setString(1, marker == null ? "" : marker);
         try (ResultSet rs = statement.executeQuery()) {
           ArrayList<Item> names = new ArrayList<>();
           while (rs.next()) {
-            names.add(
-                new Item(
-                    rs.getString(1),
-                    rs.getInt(2) == userId ? "owner" : "developer",
-                    rs.getString(3),
-                    rs.getDate(4).toString()));
+            names.add(new Item(rs.getString(1), rs.getInt(2) == userId ? "owner" : "developer", rs.getString(3), rs.getDate(4).toString()));
           }
           return names;
         }
@@ -211,14 +139,10 @@ public class Spaces {
     }
   }
 
-  public static ArrayList<String> listAllSpaceNames(DataBase dataBase)
-      throws Exception {
+  public static ArrayList<String> listAllSpaceNames(DataBase dataBase) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder("SELECT `name` FROM `")
-              .append(dataBase.databaseName) //
-              .append("`.`spaces` ORDER BY `id` ASC")
-              .toString();
+      String sql = new StringBuilder("SELECT `name` FROM `").append(dataBase.databaseName) //
+                                                            .append("`.`spaces` ORDER BY `id` ASC").toString();
       ArrayList<String> results = new ArrayList<>();
       DataBase.walk(connection, (rs) -> {
         results.add(rs.getString(1));
@@ -227,70 +151,38 @@ public class Spaces {
     }
   }
 
-  public static boolean changePrimaryOwner(
-      DataBase dataBase, int spaceId, int oldOwner, int newOwner) throws Exception {
+  public static boolean changePrimaryOwner(DataBase dataBase, int spaceId, int oldOwner, int newOwner) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("UPDATE `")
-              .append(dataBase.databaseName)
-              .append("`.`spaces` SET `owner`=")
-              .append(newOwner)
-              .append(" WHERE `id`=")
-              .append(spaceId)
-              .append(" AND `owner`=")
-              .append(oldOwner)
-              .append(" LIMIT 1")
-              .toString();
+      String sql = new StringBuilder().append("UPDATE `").append(dataBase.databaseName).append("`.`spaces` SET `owner`=").append(newOwner).append(" WHERE `id`=").append(spaceId).append(" AND `owner`=").append(oldOwner).append(" LIMIT 1").toString();
       return DataBase.executeUpdate(connection, sql) > 0;
     }
   }
 
   public static boolean delete(DataBase dataBase, int spaceId, int owner) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
-      String sql =
-          new StringBuilder()
-              .append("DELETE FROM `")
-              .append(dataBase.databaseName)
-              .append("`.`spaces` WHERE `id`=")
-              .append(spaceId)
-              .append(" AND `owner`=")
-              .append(owner)
-              .toString();
+      String sql = new StringBuilder().append("DELETE FROM `").append(dataBase.databaseName).append("`.`spaces` WHERE `id`=").append(spaceId).append(" AND `owner`=").append(owner).toString();
       return DataBase.executeUpdate(connection, sql) > 0;
     }
   }
 
-  public static void setRole(DataBase dataBase, int spaceId, int userId, Role role)
-      throws Exception {
+  public static void setRole(DataBase dataBase, int spaceId, int userId, Role role) throws Exception {
     try (Connection connection = dataBase.pool.getConnection()) {
       {
-        DataBase.execute(
-            connection,
-            new StringBuilder()
-                .append("DELETE FROM `")
-                .append(dataBase.databaseName)
-                .append("`.`grants` WHERE `space`=")
-                .append(spaceId)
-                .append(" AND `user`=")
-                .append(userId)
-                .toString());
+        DataBase.execute(connection, new StringBuilder().append("DELETE FROM `").append(dataBase.databaseName).append("`.`grants` WHERE `space`=").append(spaceId).append(" AND `user`=").append(userId).toString());
       }
       if (role != Role.None) {
-        DataBase.execute(
-            connection,
-            new StringBuilder()
-                .append("INSERT INTO `")
-                .append(dataBase.databaseName)
-                .append("`.`grants` (`space`, `user`, `role`) VALUES (")
-                .append(spaceId)
-                .append(",")
-                .append(userId)
-                .append(",")
-                .append(role.role)
-                .append(")")
-                .toString());
+        DataBase.execute(connection, new StringBuilder().append("INSERT INTO `").append(dataBase.databaseName).append("`.`grants` (`space`, `user`, `role`) VALUES (").append(spaceId).append(",").append(userId).append(",").append(role.role).append(")").toString());
       }
+    }
+  }
+
+  public static class InternalDeploymentPlan {
+    public final String plan;
+    public final String hash;
+
+    public InternalDeploymentPlan(String plan, String hash) {
+      this.plan = plan;
+      this.hash = hash;
     }
   }
